requirements.txt
streamlit~=1.41.1
numpy~=2.2.0
pillow~=10.4.0
pyzbar~=0.1.9